ci: comprehensive CI pipeline with lint, tidy, coverage, generated-test, and e2e harness
Split monolithic e2e tests into respective packages (command, agent) and
keep only the binary-level test behind //go:build e2e with a dedicated
harness. Introduce a full CI job DAG: lint (golangci-lint) and tidy
(go mod tidy check) run independently as fast gates, unit tests with
race+coverage depend on tidy, generated_templates tests run go generate
then validate the template pipeline, e2e builds the binary and exercises
it, and goreleaser snapshot validates cross-compilation.

Also fix the stale go generate path in .goreleaser.yml (pkg/scanner/resources
was renamed to pkg/tools/resources).
